+++ src/events/ready.ts
@@ -1,0 +1,3 @@
+module.exports = async (client: any) => {
+    console.log("The Bot is ready")
+}