+++ CHANGELOG.md
@@ -5,4 +5,21 @@ Listed below are the changes made to the Finnick bot.
 
 ### Added
 
-- Inital Files
+- Mongoose package
+- Discord.js package
+- Dotenv package
+- Added FClient with an Collection of Commands, Events and Schedules for future features
+- Command Handler
+- Event Handler
+
+
+## Version 0.0.0 - 12/7/20
+
+### Added
+
+- Inital files generated from `npm init`
+- Inital files generated from `tsc --init`
+- Inital files generated from `git init`
+- .gitignore
+- LICENSE
+- README